Moved extrapolation function out of estimate filled posts and into new file in job role models folder
